pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "kluris"
version = "2.14.0"
description = "Turn your AI agents into team subject matter experts with shared, human-curated knowledge"
readme = "README.md"
requires-python = ">=3.10"
license = "MIT"
dependencies = [
    "click>=8.1",
    "rich>=13.0",
    "pyyaml>=6.0",
    "python-frontmatter>=1.1",
    "pydantic>=2.0",
    "pathspec>=0.12",
]

[project.scripts]
kluris = "kluris.cli:cli"

[project.urls]
Homepage = "https://kluris.io"
Repository = "https://github.com/ngvoicu/kluris"
Issues = "https://github.com/ngvoicu/kluris/issues"

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-cov>=5.0",
    # Pack server deps — needed for tests in tests/pack/, NOT runtime deps.
    # The Docker image installs its own pinned versions of these.
    "fastapi>=0.115",
    "uvicorn[standard]>=0.30",
    "jinja2>=3.1",
    "sse-starlette>=2.1",
    "httpx>=0.27",
    "respx>=0.21",
    "pytest-asyncio>=0.23",
]

[tool.hatch.build.targets.wheel]
packages = ["src/kluris", "src/kluris_runtime"]
include = ["packaging/*"]

[tool.pytest.ini_options]
markers = [
    "llm_eval: opt-in live-model evals; require KLURIS_EVAL_* env vars",
    "docker_network: opt-in Docker e2e tests; need a Docker daemon",
]
